Use QSplitter to create splitview of trees
