+++ som_gui/module/filter_window/ui.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
 
 import logging
-from typing import Callable
+from typing import Callable,TYPE_CHECKING
+
 
 from PySide6.QtCore import (
     QAbstractItemModel,
@@ -21,7 +22,8 @@ import SOMcreator
 import som_gui
 from som_gui import tool
 from . import trigger
-
+if TYPE_CHECKING:
+    from .ui_header import CustomHeaderView
 
 class SettingsWidget(QWidget):
     def __init__(self):
@@ -38,7 +40,7 @@ class FilterWidget(QWidget):
         super().__init__(*args, **kwargs)
         from .qt import ui_Widget
 
-        self.ui = ui_Widget.Ui_FilterWindow()
+        self.ui = ui_Widget.Ui_Form()
         self.ui.setupUi(self)
         self.setWindowIcon(som_gui.get_icon())
 
@@ -66,78 +68,11 @@ class ProjectView(QTableView):
         trigger.tree_mouse_release_event(index)
 
 
-class FrozeView(QTreeView):
-    def __init__(self, parent: FilterTreeView, column_count=2):
-        super().__init__(parent)
-        self.frozen_cols = column_count
-        self.setHorizontalScrollBarPolicy(Qt.ScrollBarPolicy.ScrollBarAlwaysOff)
-        self.setVerticalScrollBarPolicy(Qt.ScrollBarPolicy.ScrollBarAlwaysOff)
-        self.setStyleSheet("""border: none""")
-
-    def update_selection(self, selected, deselected):
-        self.selectionModel().select(selected, QItemSelectionModel.SelectionFlag.Select)
-        self.selectionModel().select(
-            deselected, QItemSelectionModel.SelectionFlag.Deselect
-        )
 
 
 class FilterTreeView(QTreeView):
     def __init__(self, parent, frozen_col_count):
         super().__init__(parent)
-        self.frozen_view = FrozeView(self, frozen_col_count)
-        self.viewport().stackUnder(self.frozen_view)
-        self.header().sectionResized.connect(self.update_section_width)
-        self.frozen_view.verticalScrollBar().valueChanged.connect(
-            self.verticalScrollBar().setValue
-        )
-        self.verticalScrollBar().valueChanged.connect(
-            self.frozen_view.verticalScrollBar().setValue
-        )
-        self.frozen_view.expanded.connect(self.expand)
-        self.frozen_view.collapsed.connect(self.collapse)
-        self.expanded.connect(self.frozen_view.expand)
-        self.collapsed.connect(self.frozen_view.collapse)
-
-    def update_section_width(self, logical_index, old_size, new_size):
-        if logical_index < self.frozen_view.frozen_cols:
-            self.frozen_view.setColumnWidth(logical_index, new_size)
-            self.updateFrozenTableGeometry()
-
-    def resizeEvent(self, event):
-        super().resizeEvent(event)
-        self.updateFrozenTableGeometry()
-
-    def scrollTo(self, index, hint):
-        if index.column() >= self.frozen_view.frozen_cols:
-            super().scrollTo(index, hint)
-        else:
-            self.frozen_view.scrollTo(index, hint)
-
-    def updateFrozenTableGeometry(self):
-        x = self.frameWidth() - 1
-        y = self.frameWidth()
-        width = sum([self.columnWidth(i) for i in range(self.frozen_view.frozen_cols)])
-        height = self.viewport().height() + self.header().height()
-        self.frozen_view.setGeometry(x, y, width, height)
-        fh = self.frozen_view.header()
-        fhg = fh.geometry()
-
-        fh.setGeometry(fhg.x(), fhg.y(), fhg.width(), self.header().height())
-
-    def update_selection(self, selected, deselected):
-        self.selectionModel().select(selected, QItemSelectionModel.SelectionFlag.Select)
-        self.selectionModel().select(
-            deselected, QItemSelectionModel.SelectionFlag.Deselect
-        )
-
-    def setModel(self, model: ClassModel | PsetModel):
-        super().setModel(model)
-        self.frozen_view.setModel(model)
-        for i in range(model.columnCount()):
-            if i >= self.frozen_view.frozen_cols:
-                self.frozen_view.hideColumn(i)
-
-        self.updateFrozenTableGeometry()
 
     def mouseMoveEvent(self, event: QMouseEvent):
         super().mouseMoveEvent(event)
@@ -150,7 +85,7 @@ class FilterTreeView(QTreeView):
 
 
 class ClassTreeView(FilterTreeView):
-    def __init__(self, parent):
+    def __init__(self, parent=None):
         super().__init__(parent, 2)
 
     def enterEvent(self, event):
@@ -164,7 +99,7 @@ class ClassTreeView(FilterTreeView):
         return super().header()
     
 class PsetTreeView(FilterTreeView):
-    def __init__(self, parent):
+    def __init__(self, parent=None):
         super().__init__(parent, 1)
 
     def enterEvent(self, event):
@@ -523,433 +458,6 @@ class PsetModel(TreeModel):
         getter_funcs = [lambda o: getattr(o, "name")]
         return super().data(index, role, getter_funcs)
 
-class CustomHeaderView(QHeaderView):
-    sectionPressed = Signal(int,int)
-    def __init__(
-        self, proj:SOMcreator.SOMProject, parent=None
-    ):
-        super().__init__(Qt.Orientation.Horizontal,parent)
-        baseSectionSize = QSize()
-        baseSectionSize.setWidth(self.defaultSectionSize())
-        baseSectionSize.setHeight(20)
-        model = CustomHeaderModel(proj)
-        for row in range(model.rowCount()):
-            for col in range(model.columnCount()):
-                index = model.index(row,col)
-                model.setData(index, baseSectionSize, Qt.ItemDataRole.SizeHintRole)
-        self.setModel(model)
-        self.sectionResized.connect(self.onSectionResized)
-
-    def setModel(self,model):
-        super().setModel(model)
-
-    def setCellLabel(self, row: int, column: int, label: str):
-        self.model().setData(
-            self.model().index(row, column), label, Qt.ItemDataRole.DisplayRole
-        )
-
-    def setCellBackgroundColor(self, row: int, column: int, color: QColor):
-        self.model().setData(
-            self.model().index(row, column), color, Qt.ItemDataRole.BackgroundRole
-        )
-
-    def setCellForegroundColor(self, row: int, column: int, color: QColor):
-        self.model().setData(
-            self.model().index(row, column), color, Qt.ItemDataRole.ForegroundRole
-        )
-
-    def indexAt(self, pos: QPoint):
-        tblModel = self.model()
-        rows = tblModel.rowCount()
-        cols = tblModel.columnCount()
-        logicalIdx = self.logicalIndexAt(pos)
-        delta = 0
-        if self.orientation() == Qt.Orientation.Horizontal:
-            for row in range(rows):
-                cellIndex = tblModel.index(row, logicalIdx)
-                delta += cellIndex.data(Qt.ItemDataRole.SizeHintRole).height()
-                if pos.y() <= delta:
-                    return cellIndex
-        else:
-            for col in range(cols):
-                cellIndex = tblModel.index(logicalIdx, col)
-                delta += cellIndex.data(Qt.ItemDataRole.SizeHintRole).width()
-                if pos.x() <= delta:
-                    return cellIndex
-        return QModelIndex()
-
-    def paintSection(self, painter: QPainter, rect: QRect, logicalIndex: int):
-        tblModel:CustomHeaderModel = self.model()
-        for i in range(tblModel.rowCount()):
-            cellIndex = tblModel.index(i, logicalIndex)
-            cellSize: QSize = cellIndex.data(Qt.ItemDataRole.SizeHintRole)
-            sectionRect = QRect(rect)
-
-            sectionRect.setTop(self.rowSpanSize(logicalIndex, 0, i))
-            sectionRect.setSize(cellSize)
-
-            colSpanIdx: QModelIndex = self.columnSpanIndex(cellIndex)
-            rowSpanIdx: QModelIndex = self.rowSpanIndex(cellIndex)
-            if colSpanIdx.isValid():
-                colSpanFrom: int = colSpanIdx.column()
-                colSpanCnt = int(colSpanIdx.data(CustomHeaderModel.ColumnSpanRole))
-                colSpanTo = colSpanFrom + colSpanCnt - 1
-                colSpan = self.columnSpanSize(cellIndex.row(), colSpanFrom, colSpanCnt)
-                sectionRect.setLeft(self.sectionViewportPosition(colSpanFrom))
-                sectionRect.setWidth(colSpan)
-                subRowSpanData = colSpanIdx.data(CustomHeaderModel.RowSpanRole)
-                if subRowSpanData is not None:
-                    subRowSpanFrom = colSpanIdx.row()
-                    subRowSpanCnt = int(subRowSpanData)
-                    subRowSpanTo = subRowSpanFrom + subRowSpanCnt - 1
-                    subRowSpan = self.rowSpanSize(
-                        colSpanFrom, subRowSpanFrom, subRowSpanCnt
-                    )
-                    sectionRect.setTop(                        self.rowSpanSize(colSpanFrom, 0, subRowSpanFrom)                    )
-                    i = subRowSpanTo
-                    sectionRect.setHeight(subRowSpan)
-                cellIndex = colSpanIdx
-
-            if rowSpanIdx.isValid():
-                rowSpanFrom = rowSpanIdx.row()
-                rowSpanCnt = int(rowSpanIdx.data(CustomHeaderModel.RowSpanRole))
-                rowSpanTo = rowSpanFrom + rowSpanCnt - 1
-                rowSpan = self.rowSpanSize(cellIndex.column(), rowSpanFrom, rowSpanCnt)
-                sectionRect.setTop(self.rowSpanSize(logicalIndex, 0, rowSpanFrom))
-                i = rowSpanTo
-
-                sectionRect.setHeight(rowSpan)
-
-                # check up if the row span index has column span
-                subColSpanData = rowSpanIdx.data(CustomHeaderModel.ColumnSpanRole)
-                if subColSpanData is not None:
-                    subColSpanFrom = rowSpanIdx.column()
-                    subColSpanCnt = int(subColSpanData)
-                    subColSpanTo = subColSpanFrom + subColSpanCnt - 1
-                    subColSpan = self.columnSpanSize(
-                        rowSpanFrom, subColSpanFrom, subColSpanCnt
-                    )
-                    sectionRect.setLeft(
-                        self.sectionViewportPosition(subColSpanFrom)
-                    )
-                   
-                    sectionRect.setWidth(subColSpan)
-                cellIndex = rowSpanIdx
-
-            opt = QStyleOptionHeader()
-            self.initStyleOption(opt)
-            opt.textAlignment = Qt.AlignmentFlag.AlignCenter
-            opt.iconAlignment = Qt.AlignmentFlag.AlignVCenter
-            opt.section = logicalIndex
-            opt.text = cellIndex.data(Qt.ItemDataRole.DisplayRole)
-            if i>0:
-                pass
-                sectionRect.setTop(sectionRect.top())
-                sectionRect.setBottom(sectionRect.bottom())#-30)
-            opt.rect = sectionRect
-
-            bg = cellIndex.data(Qt.ItemDataRole.BackgroundRole)
-            fg = cellIndex.data(Qt.ItemDataRole.ForegroundRole)
-            if bg is not None:
-                opt.palette.setBrush(QPalette.ColorRole.Button, QBrush(bg))
-                opt.palette.setBrush(QPalette.ColorRole.Window, QBrush(bg))
-
-            if fg is not None:
-                opt.palette.setBrush(QPalette.ButtonText, QBrush(fg))
-
-            painter.save()
-            self.style().drawControl(QStyle.CE_Header, opt, painter, self)
-            painter.restore()
 
 
-    def sectionSizeFromContents(self, logicalIndex: int):
-        tblModel = self.model()
-        level = tblModel.rowCount()
-
-        size = super().sectionSizeFromContents(logicalIndex)
-        height = 0
-        for i in range(level):
-            cellIndex = tblModel.index(i, logicalIndex)
-            colSpanIdx: QModelIndex = self.columnSpanIndex(cellIndex)
-            rowSpanIdx: QModelIndex = self.rowSpanIndex(cellIndex)
-            size: QSize = cellIndex.data(Qt.ItemDataRole.SizeHintRole)
-            if colSpanIdx.isValid():
-                colSpanFrom = colSpanIdx.column()
-                colSpanCnt = int(colSpanIdx.data(CustomHeaderModel.ColumnSpanRole))
-                colSpanTo = colSpanFrom + colSpanCnt - 1
-                size.setWidth(
-                    self.columnSpanSize(colSpanIdx.row(), colSpanFrom, colSpanCnt)
-                )
-
-            height+= cellIndex.data(Qt.ItemDataRole.SizeHintRole).height()
-        size.setHeight(height)
-        return size
-
-    def setRowHeight(self, row: int, height: int):
-        m = self.model()
-        cols = m.columnCount()
-        for col in range(cols):
-            sz: QSize = m.index(row, col).data(Qt.ItemDataRole.SizeHintRole)
-            sz.setHeight(height)
-            m.setData(m.index(row, col), sz, Qt.ItemDataRole.SizeHintRole)
-
-    def model(self) -> CustomHeaderModel:
-        return super().model()
-
-    def setColumnWidth(self, col: int, width: int):
-        m = self.model()
-        rows = m.rowCount()
-        for row in range(rows):
-            sz: QSize = m.index(row, col).data(Qt.ItemDataRole.SizeHintRole)
-            sz.setWidth(width)
-            m.setData(m.index(row, col), sz, Qt.ItemDataRole.SizeHintRole)
-        self.resizeSection(col, width)
-
-    def setSpan(
-        self,
-        row: int,
-        column: int,
-        rowSpanCount: int = None,
-        columnSpanCount: int = None,
-    ):
-        md = self.model()
-        if rowSpanCount is None:
-            rowSpanCount = md.rowCount() 
-        if  columnSpanCount is None:
-            columnSpanCount =              1
-        idx = md.index(row, column)
-        if rowSpanCount > 0:
-            md.setData(idx, rowSpanCount, CustomHeaderModel.RowSpanRole)
-        if columnSpanCount > 0:
-            md.setData(idx, columnSpanCount, CustomHeaderModel.ColumnSpanRole)
-
-    def columnSpanIndex(self, index: QModelIndex) -> QModelIndex:
-        tblModel = self.model()
-        curRow = index.row()
-        curCol = index.column()
-        i = curCol
-        while i >= 0:
-            spanIndex = tblModel.index(curRow, i)
-            span = spanIndex.data(CustomHeaderModel.ColumnSpanRole)
-            if span is not None and spanIndex.column() + int(span) - 1 >= curCol:
-                return spanIndex
-            i -= 1
-        return QModelIndex()
-
-    def rowSpanIndex(self, index: QModelIndex) -> QModelIndex:
-        tblModel = self.model()
-        curRow = index.row()
-        curCol = index.column()
-        i = curRow
-        while i >= 0:
-            spanIndex = tblModel.index(i, curCol)
-            span = spanIndex.data(CustomHeaderModel.RowSpanRole)
-            if span is not None and spanIndex.row() + int(span) - 1 >= curRow:
-                return spanIndex
-            i -= 1
-        return QModelIndex()
-
-    def columnSpanSize(self, row: int, from_: int, spanCount: int) -> int:
-        tblModel = self.model()
-        span = 0
-        for i in range(from_,from_ + spanCount):
-            cellSize = tblModel.index(row, i).data(Qt.ItemDataRole.SizeHintRole)
-            span += cellSize.width() if cellSize else 200
-        return span
-
-    def rowSpanSize(self, column, from_, spanCount) -> int:
-        tblModel = self.model()
-        span = 0
-        for i in range(from_,from_ + spanCount):
-            cellSize = tblModel.index(i, column).data(Qt.ItemDataRole.SizeHintRole)
-            span += cellSize.height()
-        return span
-
-    def getSectionRange(self, index: QModelIndex, beginSection: int, endSection: int):
-        # TODO fix output of beginSection and endSection
-        colSpanIdx = self.columnSpanIndex(index)
-        rowSpanIdx = self.rowSpanIndex(index)
-        if colSpanIdx.isValid():
-            colSpanFrom = colSpanIdx.column()
-            colSpanCnt = int(colSpanIdx.data(CustomHeaderModel.ColumnSpanRole))
-
-            colSpanTo = colSpanFrom + colSpanCnt - 1
-            if self.orientation() == Qt.Orientation.Horizontal:
-                beginSection = colSpanFrom
-                endSection = colSpanTo
-                index = colSpanIdx
-                return colSpanCnt,beginSection,endSection
-            else:
-                subRowSpanData = colSpanIdx.data(CustomHeaderModel.RowSpanRole)
-                if subRowSpanData is not None:
-                    subRowSpanFrom = colSpanIdx.row()
-                    subRowSpanCnt = int(subRowSpanData)
-                    subRowSpanTo = subRowSpanFrom + subRowSpanCnt - 1
-                    beginSection = subRowSpanFrom
-                    endSection = subRowSpanTo
-                    index = colSpanIdx
-                    return subRowSpanCnt,beginSection,endSection
-        if rowSpanIdx.isValid():
-            rowSpanFrom = rowSpanIdx.row()
-            rowSpanCnt = int(rowSpanIdx.data(CustomHeaderModel.RowSpanRole))
-            rowSpanTo = rowSpanFrom + rowSpanCnt - 1
-            subColSpanData = rowSpanIdx.data(CustomHeaderModel.ColumnSpanRole)
-            if subColSpanData is not None:
-                subColSpanFrom = rowSpanIdx.column()
-                subColSpanCnt = int(subColSpanData)
-                subColSpanTo = subColSpanFrom + subColSpanCnt - 1
-                beginSection = subColSpanFrom
-                endSection = subColSpanTo
-                index = rowSpanIdx
-                return subColSpanCnt,beginSection,endSection
-        return 0, beginSection, endSection
-
-    def mousePressEvent(self, event:QMouseEvent):
-        super().mousePressEvent(event)
-        index = self.indexAt(event.pos())
-        if index.isValid():
-            beginSection = index.column()
-            endSection = beginSection
-            result,beginSection,endSection = self.getSectionRange(index,beginSection,endSection)
-            if result >0:
-                self.sectionPressed.emit(beginSection,endSection)
-            else:
-                self.sectionPressed.emit(beginSection,endSection)
-
-    def onSectionResized(self,logicalIndex:int,oldSize:int,newSize:int):
-        tblModel:CustomHeaderModel = self.model()
-        level =             tblModel.rowCount()   
-        pos = self.sectionViewportPosition(logicalIndex)
-        xx = pos
-        yy = 0
-        sectionRect = QRect(xx,yy,0,0)
-        for i in range(level):
-            cellIndex = tblModel.index(i,logicalIndex)
-            cellSize:QSize = cellIndex.data(Qt.ItemDataRole.SizeHintRole)
-            sectionRect.setTop(self.rowSpanSize(logicalIndex,0,i))
-            cellSize.setWidth(newSize)
-           
-            tblModel.setData(cellIndex,cellSize,Qt.ItemDataRole.SizeHintRole)
-            colSpanIdx = self.columnSpanIndex(cellIndex)
-            rowSpanIdx = self.rowSpanIndex(cellIndex)
-            if colSpanIdx.isValid():
-                colSpanFrom = colSpanIdx.column()
-                sectionRect.setLeft(self.sectionViewportPosition(colSpanFrom))
-                
-            if rowSpanIdx.isValid():
-                rowSpanFrom = rowSpanIdx.row()
-                sectionRect.setTop(self.rowSpanSize(logicalIndex,0,rowSpanFrom))  
-            rToUpdate = QRect(sectionRect)
-            rToUpdate.setWidth(self.viewport().width() - sectionRect.left())
-            rToUpdate.setHeight(self.viewport().height() - sectionRect.top())
-            print(rToUpdate,rToUpdate.normalized())
-            self.viewport().update(rToUpdate.normalized())
-
-class CustomHeaderModel(QAbstractTableModel):
-    ColumnSpanRole = Qt.ItemDataRole.UserRole + 1
-    RowSpanRole  = Qt.ItemDataRole.UserRole+2
-
-    def __init__(self,proj:SOMcreator.SOMProject,parent = None):
-        super().__init__(parent)
-        self.proj = proj
-        self.data_dict = dict()
-        self.size_hint_dict = dict()
-
-    def get_active_phases(self,usecase):
-        phases = self.proj.get_phases()
-        return [phase for phase in phases if self.proj.get_filter_state(phase,usecase)]
-
-    def get_usecase_matrix(self):
-        matrix = list()
-        for usecase in self.proj.get_usecases():
-            active_phases = self.get_active_phases(usecase)
-            if active_phases:
-                matrix.append((usecase,active_phases))
-        return matrix
-
-    def get_usecase_by_column(self,column:int):
-        column-=2
-        matrix = self.get_usecase_matrix()
-        for use_case,phase_list in matrix:
-            column -=len(phase_list)
-            if column <0:
-                return use_case
-    def get_phase_by_column(self,column:int):
-        column-=2
-        matrix = self.get_usecase_matrix()
-        for use_case,phase_list in matrix:
-            if 0<=column < len(phase_list):
-                return phase_list[column]
-            column-=len(phase_list)
-
-    def index(self,row:int,column:int,parent:QModelIndex = QModelIndex()):
-        if column <2:
-            return self.createIndex(row,column,None)
-        if row == 0:
-            return self.createIndex(row,column,self.get_usecase_by_column(column))
-        if row == 1:
-            return self.createIndex(row,column,self.get_phase_by_column(column))
-        return self.createIndex(row,column,None)
     
-    def rowCount(self,parent = None):
-        return 2
-    
-    def columnCount(self, parent = None):
-        matrix = self.get_usecase_matrix()
-        return sum(len(p) for (u,p) in matrix)+2
-    
-    def flags(self,index:QModelIndex):
-        if not index.isValid():
-            return Qt.ItemFlag.NoItemFlags
-        return super().flags(index)
-    
-    def data(self,index:QModelIndex,role:int):
-        if not index.isValid():
-            return None
-        if index.row() >= self.rowCount() or index.row() < 0 or index.column() >= self.columnCount() or         index.column() < 0:
-            return None
-        if role == Qt.ItemDataRole.SizeHintRole:
-            return self.size_hint_dict[(index.row(),index.column())]
-        if index.column() <2:
-            if role == self.ColumnSpanRole:
-                return 1
-            if role == self.RowSpanRole and index.row() == 0:
-                return 2
-            if role == Qt.ItemDataRole.DisplayRole:
-                if index.column() == 0:
-                    return "Class"
-                return "Identifier"
-            return None
-
-        item:SOMcreator.UseCase|SOMcreator.Phase = index.internalPointer()
-        sibling = index.siblingAtColumn(index.column()-1) or QModelIndex()
-        if role == self.ColumnSpanRole:
-            if sibling.internalPointer() == item:
-                return None
-            if index.row() == 1:
-                return 1
-            else:
-                return len(self.get_active_phases(item))
-        if role == self.RowSpanRole:
-            return None
-        if role == Qt.ItemDataRole.DisplayRole:
-            return str(item.name)
-
-
-        if not self.data_dict.get(item):
-                self.data_dict[item] = dict()
-        return self.data_dict.get(item).get(role)
-
-    def setData(self, index:QModelIndex, value, role):
-        if index is not None:
-            item:SOMcreator.UseCase|SOMcreator.Phase = index.internalPointer()
-            if role == self.ColumnSpanRole:
-                pass
-            if role == Qt.ItemDataRole.SizeHintRole:
-                self.size_hint_dict[(index.row(),index.column())] = value
-            else:
-                if not self.data_dict.get(item):
-                    self.data_dict[item] = dict()
-                self.data_dict[item][role] = value
-            return True
-        return False
